requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.7.4
anyio==4.4.0
astor==0.8.1
asynch==0.2.4
attrs==24.2.0
beanie==1.25.0
black==23.11.0
certifi==2024.7.4
charset-normalizer==3.3.2
ciso8601==2.3.1
click==8.1.6
clickhouse-connect==0.7.19
clickhouse-driver==0.2.9
clickhouse-sqlalchemy==0.3.2
dnspython==2.6.1
fastapi==0.95.1
frozenlist==1.4.1
h11==0.14.0
idna==3.8
injector==0.22.0
lazy-model==0.2.0
leb128==1.0.8
lz4==4.3.3
Mako==1.3.5
MarkupSafe==2.1.5
motor==3.5.1
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pydantic==1.10.14
pymongo==4.8.0
pynest-api==0.1.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.19
starlette==0.26.1
toml==0.10.2
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.23.1
yarl==1.9.4
zstandard==0.23.0
zstd==1.5.5.1
